+++ index.html
@@ -228,7 +228,7 @@
     </div>
   </template>
 
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 
 </body>
 </html>
