+++ js/util.js
@@ -1,0 +1,35 @@
+const getRandomNumber = function(min, max) {
+  if (min < 0 || max <= min) {
+    throw new Error('Неверные данные')
+  }
+
+  return Math.random() * (max - min + 1) + min;
+}
+
+//Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+
+
+const getRandomInteger = function(min, max) {
+  return Math.floor(getRandomNumber(min, max));
+}
+
+const getRandomFixed = function (min, max, fixedNumber) {
+  const rand = +(getRandomNumber(min, max));
+
+  return rand.toFixed(fixedNumber);
+}
+
+const getRandomArrayElement = (elements) => {
+  return elements[getRandomInteger(0, elements.length - 1)];
+};
+
+const getNewArray = (array, randomNumber) => {
+  const someArray = [];
+  for (let i = 0; i <= randomNumber; i++) {
+    someArray.push(array[i]);
+  }
+  const newArray = new Set(someArray);
+  return Array.from(newArray);
+};
+
+export { getRandomNumber, getRandomInteger, getRandomFixed, getRandomArrayElement, getNewArray };
